Learn about doubly linked lists and prep for chlgs
This commit is a simple commit that preps for douby linked lists
challenges by creating a new node type for doubly linked lists. The
difference between this and the other Node type is that the DNode now
includes a previous element attribute to keep track of previous nodes.

This commit also renames the node.py file to nodes.py since we are now
working with two different types of nodes in order to work with singly
and doubly linked lists.